refactor(sections): move banners into sections/banners/
Renames AppHealthBanner → HealthBanner and co-locates both banner
components under web/src/sections/banners/.
