+++ web/src/app/layout.tsx
@@ -17,8 +17,8 @@ import { ThemeProvider } from "next-themes";
 import { TooltipProvider } from "@radix-ui/react-tooltip";
 import StatsOverlayLoader from "@/components/dev/StatsOverlayLoader";
 import { cn } from "@opal/utils";
-import AppHealthBanner from "@/sections/AppHealthBanner";
-import LicenseExpiryBanner from "@/sections/LicenseExpiryBanner";
+import AppHealthBanner from "@/sections/banners/HealthBanner";
+import LicenseExpiryBanner from "@/sections/banners/LicenseExpiryBanner";
 import ProductGatingWrapper from "@/providers/ProductGatingWrapper";
 import SWRConfigProvider from "@/providers/SWRConfigProvider";
