+++ web/src/sections/banners/LicenseExpiryBanner.tsx
